Setup TopBar component
